pyproject.toml
[project]
name = "awsbot-cli"
version = "0.1.0"
description = "Agentic enabled tool that automates tasks for AWS engineers."
authors = [
    {name = "David Reay", email = "david@awsbot.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "google-genai",
    "python-gitlab (>=7.1.0,<8.0.0)",
    "jira[cli] (>=3.10.5,<4.0.0)",
    "boto3 (>=1.42.35,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.48.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.4,<2.0.0)",
    "typer (>=0.21.1,<0.22.0)",
    "gemini-cli (>=0.2.5,<0.3.0)",
    "cryptography (>=46.0.4,<47.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
awsbot-cli = "awsbot_cli.main:main"
[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-env (>=1.3.2,<2.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "rich (>=13.7.0,<14.0.0)",
    "moto (>=5.1.21,<6.0.0)"
]

[tool.pytest.ini_options]
markers = [
    "unit: mark a test as a unit test",
    "login: tests login features",
    "configure: tests configure features",
    "billing",
    "reporting"
]

[tool.ruff]
line-length = 150

[tool.pylint."FORMAT"]
max-line-length = 150

[tool.flake8]
max-line-length = 150
